CelsiusReadingをFloatベースに変更
GT7のテレメトリ温度がFloatで取得されるため、Double版のままだと
Float→Double変換時に小数ノイズ（例: 82.1f.toDouble() = 82.09999847412109）
が発生しJSON出力の見た目が悪化する。LMU/ACEのDouble値もFloatへ丸めて統一する
（温度表示は小数点1桁までしか使わないため、Floatの精度で十分）。

+++ core/domain/src/commonMain/kotlin/kurou/kodriver/domain/model/CelsiusReading.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 @JvmInline
 value class CelsiusReading(
-    val value: Double,
+    val value: Float,
 ) : Comparable<CelsiusReading> {
     override fun compareTo(other: CelsiusReading): Int = value.compareTo(other.value)
 }
